3.1.0 Add functions to get/set the close-on-exec flag (#34)
pid2 needs to set the close-on-exec flag on a certain file descriptor,
but node doesn't have an API to achieve this. The only other node module
that provides the `fcntl(2)` binding doesn't play well with esbuild, so
we need to recruit the help of our friend here.

This change adds two functions to get/set the close-on-exec flag on a
file descriptor.