3.1.2 Set the PTY fd to be non-blocking
We have gone back and forth a couple of times, but we _do_ need to set
this FD to be non-blocking. Otherwise node will fully consume one I/O
thread for each PTY, and node only has four of those.

This change makes the controller fd (the one the parent process reads
from / writes to) as non-blocking for node. Also, now that the
controller fd is non-blocking, we need to pass it through Node's TTY
machinery so that it is properly handled. Because of that, Rust is no
longer on the hook for closing the FD, since ownership is fully
transferred to Node.